+++ extensions/tidyread---streamline-your-daily-reading/CHANGELOG.md
@@ -1,6 +1,8 @@
 # TidyRead - Streamline Your Daily Reading Changelog
 
 ## [Improving Digest Speed] - 2024-01-26
+
+- add progress display for generating digest
 - openai api change to normal response instead of stream
 - improve translate prompt
 - change api config based on provider instead of same
@@ -8,8 +10,9 @@
 - requestTimeout default to 30s
 
 ## [Onboarding Page] - 2024-01-24
+
 - add quickly generating sources by user interest
-- add retryCount and retryDelay setting to decrease article summarizing error 
+- add retryCount and retryDelay setting to decrease article summarizing error
 - improve speed of batch import sources
 - improve error handling and error message
 - add Delete All Sources
@@ -17,4 +20,5 @@
 - fix fetch large html content memory leak
 
 ## [Initial Version] - 2024-01-06
+
 init